--- Services/Stories/Stories.Application/Commands/CreateStoryCommand.cs
@@ -1,11 +1,0 @@
-﻿using MediatR;
-using Stories.Application.Responses;
-
-namespace Stories.Application.Commands
-{
-    public class CreateStoryCommand : IRequest<CreateStoryResponse>
-    {
-        public string StoriesImageUrl { get; set; } = null!;
-        public string CreatorNickName { get; set; } = null!;
-    }
-}
