--- Services/Stories/Stories.Application/Commands/DeleteStoryCommand.cs
@@ -1,10 +1,0 @@
-﻿using MediatR;
-using Stories.Application.Responses;
-
-namespace Stories.Application.Commands
-{
-    public class DeleteStoryCommand : IRequest<DeleteStoryResponse>
-    {
-        public string StoryId { get; set; } = null!;
-    }
-}
