+++ Services/Stories/Stories.Application/Features/Stories/GetStoryByCreatorQuery.cs
@@ -1,12 +1,15 @@
 ﻿using MediatR;
 using Stories.Application.Mappers;
-using Stories.Application.Queries;
 using Stories.Application.Responses;
 using Stories.Domain.Entity;
 using Stories.Domain.Repositories;
-
-namespace Stories.Application.Handles
+namespace Stories.Application.Features.Stories
 {
+    public class GetStoryByCreatorQuery : IRequest<IList<GetStoryResponse>>
+    {
+        public string NickName { get; set; } = null!;
+    }
+
     public class GetStoryByCreatorHandler(IStoriesRepository storiesRepository) : IRequestHandler<GetStoryByCreatorQuery, IList<GetStoryResponse>>
     {
         private readonly IStoriesRepository _storiesRepository = storiesRepository;
